api/requirements.txt
amqp==5.1.1
apturl==0.5.2
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.4.0
black==23.7.0
blinker==1.4
Brlapi==0.8.3
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.3
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-database-url==2.0.0
dj-email-url==1.0.6
dj-rest-auth==4.0.0
Django==4.2.4
django-allauth==0.52.0
django-auto-prefetch==1.5.1
django-cache-url==3.4.4
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.3
django-jazzmin==2.6.0
django-model-utils==4.3.1
django-notifications-hq==1.8.2
django-redis==5.3.0
django-social-share==2.3.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
drf-yasg==1.21.5
duplicity==0.8.21
environs==9.5.0
exceptiongroup==1.1.3
executing==1.2.0
extensions==0.4
fasteners==0.14.1
filelock==3.12.2
flower==2.0.0
future==0.18.2
gpg==1.16.0
gunicorn==21.2.0
gyp==0.1
httplib2==0.20.2
humanize==4.7.0
identify==2.5.27
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.14.0
isort==5.12.0
itypes==1.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jsonfield==3.1.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.5.0
kombu==5.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
natsort==8.0.2
netifaces==0.11.0
nltk==3.8.1
nodeenv==1.8.0
numpy==1.21.5
oauthlib==3.2.2
olefile==0.46
packaging==23.1
paramiko==2.9.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.4.13
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.3.3
prettytable==3.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==3.12.4
psycopg2==2.9.6
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.8.0
pylint==2.17.5
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.4.0
python-apt==2.4.0+ubuntu2
python-crontab==3.0.0
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3
pyxdg==0.27
PyYAML==6.0.1
redis==4.5.5
referencing==0.30.2
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
starkbank-ecdsa==2.0.3
swapper==1.3.0
systemd-python==234
texttable==1.6.4
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.4
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.24.3
wadllib==1.3.6
wcwidth==0.2.6
websocket-client==1.2.3
whitenoise==6.1.0
wrapt==1.15.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
